Move python-version specific code to compat.py
